+++ Software/MATLAB/system/blocks/sf_decode_flat_json_object_cb.m
@@ -96,7 +96,4 @@ function idx = findParamIndex(blk, name)
 
 
 end
-%
-% function toggleSettings(blk)
-%
-% end
+